Finish up that change to 32bits
Also don't do that 'preliminary' thing, it fucks up the raw flag and
isn't needed in the first place.

+++ src/bef.h
@@ -253,6 +253,6 @@ void *bef_realloc(void *ptr, size_t sz);
 void *bef_reallocarray(void *ptr, size_t nmemb, size_t sz);
 
 /* Get max number of fragments for a given parity type */
-uint16_t bef_max_frag(bef_par_t par_t);
+uint32_t bef_max_frag(bef_par_t par_t);
 
 #endif /* BEF_H */
